Moved PackParser methods to CITReloadListener
